refactor: API 폴더 구조를 도메인별 폴더로 리팩토링
- src/api/authApi.ts → src/api/auth/index.ts + types.ts
- src/api/userApi.ts → src/api/user/index.ts + types.ts
- src/api/workerApi.ts → src/api/worker/index.ts + types.ts
- src/types/api.types.ts에서 auth 전용 타입 분리, 공통 타입만 유지
- src/types/user/, src/types/worker/ → 각 API 도메인 폴더로 이동
- 불필요한 파일 삭제 (userApiRequest/Response.type.ts, workerApiRequest/Response.type.ts, signup.types.ts)
- 모든 import 경로 업데이트
