+++ src/api/user/types.ts
@@ -13,6 +13,13 @@ export interface UserResponse {
   profileImageUrl?: string;
 }
 
+/** [PUT] /api/users/me 요청 데이터 */
+export interface UserUpdateRequest {
+  name?: string;
+  phone?: string;
+  profileImageUrl?: string;
+}
+
 // ============ 근로자 정보 관련 타입 ============
 
 /** [GET] /api/workers/user/{userId}, [PUT] /api/users/me/account 응답 데이터 */
@@ -22,6 +29,12 @@ export interface WorkerResponse {
   name: string;
   phone: string;
   workerCode: string;
+  accountNumber: string;
+  bankName: string;
+}
+
+/** [PUT] /api/users/me/account 요청 데이터 (근로자 전용) */
+export interface WorkerUpdateRequest {
   accountNumber: string;
   bankName: string;
 }
