--- src/types/signup.types.ts
@@ -1,1 +1,0 @@
-export type UserType = "WORKER" | "EMPLOYER";
